[BY-254] Refactor: make authenticate folder
