+++ lib/model/authenticate/firebase_user.dart
