+++ lib/model/authenticate/nickname_controller.dart
