--- lib/model/login_user.dart
@@ -1,6 +1,0 @@
-class LoginUser {
-  final String? email;
-  final String? password;
-
-  LoginUser({this.email, this.password});
-}
